+++ mvc/src/main/java/com/interface21/webmvc/servlet/view/resolver/ViewResolver.java
@@ -1,6 +1,7 @@
-package com.interface21.webmvc.servlet;
+package com.interface21.webmvc.servlet.view.resolver;
 
 import com.interface21.webmvc.servlet.view.JspView;
+import com.interface21.webmvc.servlet.view.View;
 
 public class ViewResolver {
